+++ neopixel.h
@@ -22,9 +22,9 @@ extern "C" {
 
     void set_leds_colour(Colour* colour);
 
-    void set_leds_colour_by_angle_1(int angle);
+    void set_leds_colour_by_angle_1(uint16_t angle);
     
-    void set_leds_colour_by_angle_2(int angle);
+    void set_leds_colour_by_angle_2(uint16_t angle);
 
 #ifdef	__cplusplus
 }
